requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
python-dotenv==1.0.0
pillow==10.2.0
matplotlib==3.8.2
pandas==2.2.0
numpy==1.26.3
